+++ lib/collections/transform/cartOrder.js
@@ -1,4 +1,6 @@
 import accounting from "accounting-js";
+import _ from "lodash";
+import { Shops } from "/lib/collections";
 
 /**
  * getSummary
@@ -54,9 +56,19 @@ function getSummary(items, prop, prop2, shopId) {
  * in code: Cart.findOne().getTotal()
  */
 export const cartOrderTransform = {
+  /**
+   * Return the total quantity on the order
+   * @method orderCount
+   * @returns {Number}  Total quantity of items on the order
+   */
   getCount() {
     return getSummary(this.items, ["quantity"]);
   },
+  /**
+   * Return the total price of shipping/handling on the order
+   * @method orderShipping
+   * @returns {Number} Total price of shipping/handling on the order
+   */
   getShippingTotal() {
     // loop through the cart.shipping, sum shipments.
     const rate = getSummary(this.shipping, ["shipmentMethod", "rate"]);
@@ -64,6 +76,23 @@ export const cartOrderTransform = {
     const shipping = handling + rate || 0;
     return accounting.toFixed(shipping, 2);
   },
+  /**
+   * Get the total price of shipping, broken down by shop
+   * @method getShippingByShop
+   * @returns {{Object}} - Total price of shipping, broken down by shop
+   */
+  getShippingByShop() {
+    const billingObject = {};
+    for (const billingItem of this.billing) {
+      billingObject[billingItem.shopId] = billingItem.invoice.shipping;
+    }
+    return billingObject;
+  },
+  /**
+   * Return the total price of goods on an order
+   * @method orderSubTotal
+   * @returns {Number} Total price of goods for the order
+   */
   getSubTotal() {
     const subTotal = getSummary(this.items, ["quantity"], ["variants", "price"]);
     return accounting.toFixed(subTotal, 2);
@@ -84,7 +113,11 @@ export const cartOrderTransform = {
       return uniqueShopSubTotals;
     }, {});
   },
-
+  /**
+   * Total taxes for order
+   * @method orderTaxes
+   * @returns {Number} Total price of taxes for an order
+   */
   getTaxTotal() {
     // taxes are calculated in a Cart.after.update hooks
     // the tax value stored with the cart/order is the effective tax rate
@@ -115,11 +148,20 @@ export const cartOrderTransform = {
       return shopTaxTotals;
     }, {});
   },
-
+  /**
+   * Discount for order
+   * @method orderDiscounts
+   * @return {Number} Total value of discounts
+   */
   getDiscounts() {
     const discount = this.discount || 0;
     return accounting.toFixed(discount, 2);
   },
+  /**
+   * Total for Order
+   * @method orderTotal
+   * @return {Number} Total for order
+   */
   getTotal() {
     const subTotal = parseFloat(this.getSubTotal());
     const shipping = parseFloat(this.getShippingTotal());
@@ -129,7 +171,6 @@ export const cartOrderTransform = {
     const total = discountTotal + shipping + taxes;
     return accounting.toFixed(total, 2);
   },
-
   /**
    * Aggregates the cart/order total by shopId
    * @method getTotalByShop
@@ -166,5 +207,64 @@ export const cartOrderTransform = {
       }
       return itemsByShop;
     }, {});
+  },
+  /**
+   * Returns an array of payment methods, normalized
+   * @method getPaymentMethod
+   * @returns {Array} Array of Payment Method objects
+   */
+  getPaymentMethods() {
+    const billingMethods = this.billing.map((method) => {
+      return method.paymentMethod;
+    });
+    const methodObjects = billingMethods.map((method) => {
+      const paymentMethodObject = {
+        storedCard: method.storedCard,
+        processor: method.processor,
+        mode: method.mode,
+        transactionId: method.transactionId,
+        amount: method.amount,
+        method: method.method
+      };
+      return paymentMethodObject;
+    });
+    return methodObjects;
+  },
+  /**
+   * Create an object that gives a summary for each shop
+   * @method getShopSummary
+   * @return {Object}  An object with a key for each shopId, and name + summary data for each
+   */
+  getShopSummary() {
+    // massage items into an object by Shop
+    const taxesByShop = this.orderTaxesByShop();
+    const subTotalsByShop = this.orderSubTotalByShop();
+    const shippingByShop = this.orderShippingByShop();
+    const itemsByShop = this.items.reduce((shopItems, item) => {
+      if (!shopItems[item.shopId]) {
+        shopItems[item.shopId] = [item];
+      } else {
+        shopItems[item.shopId].push(item);
+      }
+      return shopItems;
+    }, {});
+
+    const shopObjects = Object.keys(itemsByShop).map(function (shop) {
+      return {
+        [shop]: {
+          name: Shops.findOne(shop).name,
+          subTotal: subTotalsByShop[shop],
+          taxes: taxesByShop[shop],
+          items: itemsByShop[shop],
+          quantityTotal: itemsByShop[shop].reduce((qty, item) => {
+            return qty + item.quantity;
+          }, 0),
+          shipping: shippingByShop[shop]
+        }
+      };
+    });
+
+    const sortedShopObjects = _.sortBy(shopObjects, (shopObject) => shopObject.name);
+    return sortedShopObjects;
   }
 };
